fix: modificando migraciones de tenants para validar fqdn unico

+++ database/migrations/tenant/2022_08_09_000017_create_usuarios_table.php
@@ -24,7 +24,7 @@ class CreateUsuariosTable extends Migration
       $table->string('contraseña', 50);
 
       // Llave foranea
-      $table->integer('rol_id')->unsigned();
+      $table->bigInteger('rol_id')->unsigned();
       $table->foreign('rol_id')->references('rol_id')->on('roles');
 
       $table->rememberToken();
